@@ -2,57 +2,38 @@
   "name": "michael-simoneau-com",
   "private": true,
   "version": "2.0.0",
-  "main": "expo-router/entry",
+  "type": "module",
   "scripts": {
-    "start": "expo start --web",
-    "android": "expo run:android",
-    "ios": "expo run:ios",
-    "web": "expo start --web",
-    "build": "expo export --platform web --output-dir dist --clear",
-    "test": "jest",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest run",
     "lint": "eslint .",
     "typecheck": "tsc --noEmit",
-    "clean": "rm -rf dist .expo",
-    "release": "netlify deploy --dir=dist --prod --message=\"Expo Netlify release\""
+    "format": "prettier --write ."
   },
   "dependencies": {
-    "expo": "~51.0.20",
-    "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.7",
-    "expo-status-bar": "~1.12.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.74.5",
-    "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "~3.31.1",
-    "react-native-web": "0.19.11"
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@babel/preset-env": "^7.25.4",
-    "@testing-library/dom": "^10.4.1",
-    "@testing-library/jest-native": "5.4.3",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/react-native": "12.4.0",
-    "@types/node": "^24.8.0",
-    "@types/react": "~18.2.79",
-    "@types/react-dom": "~18.2.25",
-    "babel-jest": "^29.7.0",
-    "eslint": "^9.9.1",
-    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
-    "eslint-plugin-react-refresh": "^0.4.11",
-    "expo-constants": "~16.0.2",
-    "expo-doctor": "^1.17.10",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
-    "jest-expo": "^51.0.3",
-    "jest-transform-stub": "^2.0.0",
-    "netlify-cli": "^17.33.1",
-    "prettier": "^3.3.3",
-    "react-test-renderer": "18.2.0",
-    "resize-observer-polyfill": "^1.5.1",
-    "typescript": "~5.3.3",
-    "typescript-eslint": "^8.46.1"
+    "@testing-library/dom": "10.4.0",
+    "@testing-library/jest-dom": "6.6.3",
+    "@testing-library/react": "16.0.1",
+    "@testing-library/user-event": "14.5.2",
+    "@types/node": "20.16.10",
+    "@types/react": "18.3.7",
+    "@types/react-dom": "18.3.2",
+    "@vitejs/plugin-react": "4.3.1",
+    "eslint": "9.11.1",
+    "eslint-plugin-react-hooks": "5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "0.4.12",
+    "jsdom": "25.0.1",
+    "prettier": "3.3.3",
+    "typescript": "5.5.4",
+    "typescript-eslint": "8.8.1",
+    "vite": "5.4.8",
+    "vitest": "2.1.1"
   },
-  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
+  "packageManager": "yarn@1.22.22"
 }
